src/requirements_node.txt
ansible==10.5.0
ansible-core==2.17.5
asyncio==3.4.3
asyncpg==0.30.0
cffi==1.17.1
cryptography==43.0.3
Jinja2==3.1.4
MarkupSafe==3.0.2
nats-py==2.9.0
packaging==24.1
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
resolvelib==1.0.1